# apis/backend-api/web/openapi.yaml
openapi: 3.1.2
info:
  title: SDKWork Webserver Backend API
  description: SDKWork Web Server management and control-plane API.
  version: 1.0.0
servers:
- url: https://api.server.sdkwork.com
  description: Production
- url: http://localhost:3800
  description: Local Development
security:
- AuthToken: []
  AccessToken: []
paths:
...
  /backend/v3/api/agent/heartbeat:
    post:
      operationId: heartbeat
      summary: Report an edge-agent heartbeat
      tags:
      - agent
      security:
      - AgentToken: []
        AccessToken: []
      x-sdkwork-api-surface: backend-api
      x-sdkwork-api-authority: sdkwork-webserver-backend-api
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-auth-mode: api-key
      x-sdkwork-route-auth: agent-token
      x-sdkwork-owner: sdkwork-webserver
      x-sdkwork-rate-limit-tier: auth-critical
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentHeartbeatRequest'
      responses:
        '200':
          description: Edge-agent heartbeat acknowledged successfully.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/AgentHeartbeatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters: []
...
components:
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
    AgentToken:
      type: apiKey
      in: header
      name: X-SDKWork-Agent-Token
...
  schemas:
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required:
      - type
      - title
      - status
      - code
      - traceId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          $ref: '#/components/schemas/SdkWorkPlatformErrorCode'
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...
    AgentHeartbeatRequest:
      type: object
      additionalProperties: false
      properties:
        agentVersion:
          type: string
          maxLength: 100
        nginxEnabled:
          type: boolean
        activeConfigs:
          type: string
          format: int64
          pattern: ^-?[0-9]+$
          x-sdkwork-int64-string: true
          x-sdkwork-rust-type: i64
          description: Number of active nginx configs reported by the agent as a string.
        lastSyncVersion:
          type: string
          pattern: ^sv1:[a-f0-9]{64}$
          description: Last successfully applied syncVersion reported by the edge
            agent.
        certificateObservations:
          type: array
          maxItems: 2048
          default: []
          items:
            $ref: '#/components/schemas/AgentCertificateObservation'
    AgentCertificateObservation:
      type: object
      additionalProperties: false
      required:
      - certificateId
      - fingerprint
      - syncVersion
      - state
      - observedAt
      properties:
        certificateId:
          type: string
          format: uuid
        fingerprint:
          type: string
          pattern: ^[a-f0-9]{64}$
        syncVersion:
          type: string
          pattern: ^sv1:[a-f0-9]{64}$
        state:
          type: string
          enum:
          - STAGED
          - ACTIVE
          - SERVED
          - FAILED
        observedAt:
          type: string
          format: date-time
        failureCode:
          type: string
          pattern: ^[A-Z0-9][A-Z0-9_.-]{0,63}$
    AgentHeartbeatResponse:
      type: object
      required:
      - serverId
      - status
      - acknowledgedAt
      properties:
        serverId:
          type: string
          format: uuid
        status:
          type: integer
        acknowledgedAt:
          type: string
          format: date-time
...
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - data
      - traceId
      properties:
        code:
          type: integer
          format: int32
          enum:
          - 0
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload typed per response schema.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
...
    SdkWorkPlatformErrorCode:
      type: integer
      format: int32
      minimum: 40001
      maximum: 79999
      description: Platform or domain error code per API_SPEC.md section 15.3.
    FieldError:
      type: object
      additionalProperties: false
      required:
      - field
      - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
          minimum: 40011
          maximum: 40099
...
  responses:
...
    Unauthorized:
      description: Authentication is required.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...